optimize the codeflow
