+++ tools/SampleAnalyzer/Commons/DataFormat/RecEventFormat.h
@@ -51,8 +51,7 @@ namespace MA5
 {
 
 namespace Substructure {
-    class VariableR;
-    class Cluster;
+    class ClusterBase;
 }
 class LHEReader;
 class LHCOReader;
@@ -75,8 +74,7 @@ class RecEventFormat
   friend class DelphesTreeReader;
   friend class DelphesMA5tuneTreeReader;
   friend class DelphesMemoryInterface;
-  friend class Substructure::VariableR;
-  friend class Substructure::Cluster;
+  friend class Substructure::ClusterBase;
 
   // -------------------------------------------------------------
   //                        data members
